+++ Classes/Transaction.py
@@ -5,7 +5,7 @@ class Transaction(db.Model):
     coin_name = db.Column(db.String(), db.ForeignKey('coin.name'), nullable = False)
     user_id = db.Column(db.Integer(), db.ForeignKey('user.id'), nullable = False)
     date = db.Column(db.String(), nullable = False)
-    amount = db.Column(db.Integer(), nullable=False)
+    amount = db.Column(db.Float(), nullable=False)
     price = db.Column(db.Float(), nullable = False)
